requirements.txt
requests
pywebview
lxml
beautifulsoup4
javascript
